botão filtragem funcionando console.log
